merges with debos developments
